+++ portal/users/users-crud/hard_delete_users.php
@@ -10,6 +10,11 @@ if (!isset($_SESSION['user_id'])) {
     exit();
 }
 
+if ($_SESSION["user_type"] == "editor") {
+    header("Location: ../../dashboard.php");
+    exit();
+}
+
 if (isset($_GET['id'])) {
     $id = $_GET['id'];
     $conn->query("DELETE FROM users WHERE id = $id");
